requirements.txt
streamlit
diffusers
transformers
accelerate
torch
torchvision
Pillow
